+++ CHANGELOG.md
@@ -4,12 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org).
 
-## [Unreleased]
+## [0.15.0] - 2025-06-17
 ### Added
 * *Nothing*
 
 ### Changed
-* *Nothing*
+* Update to shlink-frontend-kit 1.0.0
+* No longer use `!important` with tailwind styles
 
 ### Deprecated
 * *Nothing*